Crate statedb

StateDB stores the entire state of the chain, including current state as well as all historical ones. Starting from the genesis state S_0, each transaction TX_i bumps the version from S_i to S_{i+1}. The state at each version consists of two components:

The mapping from access paths to current values.

All the contract events emitted up to this point in time, from the beginning of history.

There is a Merkle tree that commits to both components. For example, imagine state S_1 has four access paths A, B, C, D, each of which has the value 10, and there have been three event entries. The entire Merkle tree would look like the following:

                     state root hash 1
                  /                     \
                 /                       \
                /                         \
               /                           \
           hash                             hash
        /         \                     /          \
    hash           hash             hash            hash
   /    \         /    \           /    \          /    \
  A:10   B:10 C:10      D:10   Event0     Event1 Event2      [dummy]

If the next transaction TX_1 updates A and D's values and emits two new events, the Merkle tree would become:

                               state root hash 2
                         /                           \
                       /                               \
                     /                                   hash
                   /                                 /             \
                 /                                /                   \
               /                                /                       \
           hash                             hash                         hash
        /         \                     /          \                    /    \
    hash           hash             hash            hash            hash      [dummy]
   /    \         /    \           /    \          /    \          /    \
  A:7    B:10 C:10      D:13   Event0     Event1  Event2     Event3  Event4      [dummy]

This way a given root hash not only fixes the value of each access path, it also fixes the entire set of events up to this point in time.

Modules

state_encoder
The purpose of this library is to encode StateDB data structures in a way that is suitable for storing in a key value store.

Structs

StateDB

StateDBVersionView
StateDBVersionView is like a snapshot of StateDB at a specific version. It is passed to virtual machine for contract execution, during which the vm is free to read anything at the given version. But the vm can't read anything at other versions or write to StateDB.

StateValueWithVersion

StateValueWithVersionAndProof

StorageTransaction

Constants

NUM_SHARDS

Type Definitions

ShardId

StateValue